refactor: 新增 `GunFireSystem` 用來處理開槍，移除 `GunActionSystem`
